fix(ui): extract forecast utils to separate file for react-refresh
Move parseForecastDays, CONDITION_ICONS, CONDITION_COLORS to a shared
utils file to satisfy react-refresh/only-export-components lint rule.
